pythonfsdgroup2capstone-main/django/profile-rest-api/requirements.txt
django
Djangorestframework
pymysql
